requirements.txt
langchain-astradb
langchain
langchain_huggingface
pypdf 
datasets
flask
python-dotenv

-e .